Admin routes split
